+++ N2A/src/gov/sandia/n2a/ui/eq/tree/NodeVariable.java
@@ -248,10 +248,10 @@ public class NodeVariable extends NodeContainer
         String result = source.key ();
         Variable.ParsedValue pieces = new Variable.ParsedValue (getValue ());
 
-        int offset = tabs.get (0).intValue () - fm.stringWidth (result);
+        int offset = tabs.get (0) - fm.stringWidth (result);
         result = result + pad (offset, fm) + "=" + pieces.combiner;
 
-        offset = tabs.get (1).intValue () - fm.stringWidth (result);
+        offset = tabs.get (1) - fm.stringWidth (result);
         result = result + pad (offset, fm) + pieces.expression;
         if (! pieces.condition.isEmpty ()) result = result + " @ " + pieces.condition;
